pipes: now dropping items of malformed pipes; altered behavior of require-sign
Before this change, signs have only been required at the input-piston, which - in
many real-world cases - can be quite cumbersome from the perspective of physical space
to put a sign at. Now, the behavior is as follows: once a valid pipe-sign has been
encountered, all subsequent exit-nodes become enabled. This works great for two reasons:
firstly, the input-piston doesn't need a sign since one at the next output suffices
and secondly, the overflow-chest at the very end will always be activated (except
for the base-case of a two-node pipe, which is irrelevant, as the user needs to validate
the pipe at either end to prove their permissions), which avoids unendingloops.